Moving use fetch Hooks in their own modules
