+++ app/hooks/fetch/useFetchUserProjectsWithToken.ts
@@ -1,0 +1,40 @@
+import * as React from 'react'
+import axios, {Canceler} from 'axios'
+import {useProjectContext} from '@hooks/project/useProjectContext'
+import type {ProjectState} from 'app/types/project'
+
+export default function useFetchUserProjectsWithToken(
+  token: string
+): ProjectState {
+  const {projects, persist, clear} = useProjectContext()
+  React.useEffect(() => {
+    let cancel: Canceler
+    ;(async () => {
+      try {
+        const {
+          data: {projects},
+        } = await axios.get('/project/user', {
+          headers: {
+            Authorization: `Bearer ${token}`,
+          },
+          cancelToken: new axios.CancelToken(c => (cancel = c)),
+        })
+        persist(projects)
+      } catch (thrown) {
+        if (axios.isCancel(thrown)) {
+          console.log('Request canceled', thrown.message)
+          return thrown.message
+        } else {
+          Promise.reject(thrown)
+        }
+      }
+    })()
+    //* Cleanup
+    return () => {
+      //* cancel the request
+      cancel()
+      clear()
+    }
+  }, [clear, persist, token])
+  return projects
+}
